+++ sdk/monitor/azure-monitor-opentelemetry/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features Added
 
+- Added LogRecordProcessors to allow passing custom log record processors
 - Performance Counters
   ([#43262](https://github.com/Azure/azure-sdk-for-python/pull/43262))
 - Adding more diagnostic log message IDs